+++ spec/Suite/psr11/ContainerSpec.php
@@ -1,5 +1,6 @@
 <?php declare(strict_types=1);
 
+use suda\Registry;
 use suda\psr11\Container;
 use suda\psr11\ContainerException;
 use suda\psr11\NotFoundException;
@@ -19,30 +20,17 @@ class FakeContainer implements \Psr\Container\ContainerInterface {
 
 describe('Container', function() {
     describe('__construct', function() {
-        it('accepts values', function() {
-            $di = new Container([
-                'param' => 'value',
-                stdClass::class => stdClass::class,
-            ]);
-
-            expect($di->get('param'))->toBe('value');
-            expect($di->get(stdClass::class))->toBeAnInstanceOf(stdClass::class);
-        });
-        it('accepts container as delegate', function() {
-            $delegate = new Container([RecursiveIterator::class => function() {
-                return new RecursiveArrayIterator(['/' => ['plants', 'rocks']]);
-            }]);
-
-            expect((new Container([], $delegate))->make(ParentIterator::class))->toBeAnInstanceOf(ParentIterator::class);
+        it('accepts a Registry', function() {
+            expect(new Container(new Registry))->toBeAnInstanceOf(Container::class);
         });
     });
 
-    describe('withPsrContainer', function() {
+    describe('withDelegate', function() {
         it('accepts a container as delegate', function() {
-            $di = new Container([Exception::class => ['$message']]);
-            $di = $di->withPsrContainer(new FakeContainer);
+            $container = new Container(new Registry([Exception::class => ['$message']]));
+            $container = $container->withDelegate(new FakeContainer);
 
-            $service = $di[Exception::class];
+            $service = $container->get(Exception::class);
 
             expect($service)->toBeAnInstanceOf(Exception::class);
             expect($service->getMessage())->toBe('All your base are belong to us');
@@ -51,63 +39,63 @@ describe('Container', function() {
 
     describe('get', function() {
         it('returns param', function() {
-            $di = new Container;
+            $di = new Registry;
             $di['param'] = 'value';
 
-            expect($di->get('param'))->toBe('value');
+            expect((new Container($di))->get('param'))->toBe('value');
         });
         it('returns service', function() {
-            $di = new Container;
+            $di = new Registry;
             $di[stdClass::class] = function() {
                 return new stdClass;
             };
 
-            expect($di->get(stdClass::class))->toBeAnInstanceOf(stdClass::class);
+            expect((new Container($di))->get(stdClass::class))->toBeAnInstanceOf(stdClass::class);
         });
         it('throws NotFoundException when entry is not found', function() {
             expect(function() {
-                $di = new Container;
-                $di->get('not_found');
+                $container = new Container(new Registry);
+                $container->get('not_found');
             })
             ->toThrow(new NotFoundException);
         });
         it('throws ContainerException when factory returns invalid instance', function() {
-            $di = new Container;
+            $di = new Registry;
             $di[stdClass::class] = function() {
                 return new SplFixedArray;
             };
 
             expect(function() use($di) {
-                $di->get(stdClass::class);
+                (new Container($di))->get(stdClass::class);
             })
             ->toThrow(new ContainerException);
         });
         it('throws ContainerException when abstract class being make in factory', function() {
-            $di = new Container;
+            $di = new Registry;
             $di[stdClass::class] = function($c, $make) {
                 $make(Countable::class);
             };
 
             expect(function() use($di) {
-                $di->get(stdClass::class);
+                (new Container($di))->get(stdClass::class);
             })
             ->toThrow(new ContainerException);
         });
         it('throws ContainerException when parameter is not found', function() {
-            $di = new Container;
+            $di = new Registry;
             $di[ParentIterator::class] = ParentIterator::class;
 
             expect(function() use($di) {
-                $di->get(ParentIterator::class);
+                (new Container($di))->get(ParentIterator::class);
             })
             ->toThrow(new ContainerException);
         });
         it('throws ContainerException when cyclic dependency happen', function() {
-            $di = new Container;
+            $di = new Registry;
             $di[ParentIterator::class] = [ParentIterator::class];
 
             expect(function() use($di) {
-                $di->get(ParentIterator::class);
+                (new Container($di))->get(ParentIterator::class);
             })
             ->toThrow(new ContainerException);
         });
@@ -115,10 +103,10 @@ describe('Container', function() {
 
     describe('has', function() {
         it('returns true when param exists', function() {
-            $di = new Container;
+            $di = new Registry;
             $di['param'] = 'value';
 
-            expect($di->has('param'))->toBe(true);
+            expect((new Container($di))->has('param'))->toBe(true);
         });
     });
 });